Removed interface HeatmapDAO
